Add ebpf_object_load_native_by_fds
Add a function to load a native image without going through the bpf_object
abstraction. The function expects program and map fd arrays to be allocated
by the caller because dealing with memory which is dynamically allocated in
C/C++ is quite cumbersome in Go. Returns a size hint and a well known error
in case the arrays are too small.